Resolved issues
Changes applied regarding the variable names and the description in .yaml file and the offset in the OSNAP data is fixed now. I think the problem was just, that with OSNAP we only have monthly data which was timestamped at 12pm not 12 am. this offset wasn't part of the plotting. I added function to align the monthly means with the raw data only if it's not 12 am and the issue seems resolved now.